# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.CoffeeShop:
    properties:
      address:
        type: string
      city:
        type: string
      location:
        items:
          type: number
        type: array
      name:
        type: string
      rating:
        type: number
      roaster:
        type: boolean
    type: object
  models.EmptyBody:
    type: object
...
  models.GetUserResponse:
    properties:
      bio:
        type: string
      email:
        type: string
      firstName:
        type: string
      id:
        type: string
      isStaff:
        type: string
      lastName:
        type: string
      username:
        type: string
    type: object
  models.LikeUnlikeCoffeeShopRequest:
    properties:
      shopId:
        type: string
      userId:
        type: string
    type: object
  models.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  models.LoginResponse:
    properties:
      token:
        type: string
    type: object
  models.SignUpRequest:
    properties:
      email:
        type: string
      password:
        type: string
      passwordConfirmation:
        type: string
      username:
        type: string
    type: object
  models.UpdateUserRequest:
    properties:
      bio:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      username:
        type: string
    type: object
  types.ApiError:
    properties:
      errors:
        additionalProperties:
          type: string
        type: object
      message:
        type: string
    type: object
host: go-coffee-api.vercel.app
info:
  contact:
    email: eduardozepeda@coffeebytes.dev
    name: Eduardo Zepeda
  description: This API returns information about speciality coffee shops in Guadalajara,
    Mexico.
  license:
    name: MIT
    url: https://mit-license.org/
  termsOfService: http://swagger.io/terms/
  title: Coffee Shops in Gdl API
  version: "1.0"
paths:
...
  /likes:
    get:
      consumes:
      - application/json
      description: Return liked coffee shops data by user id
      parameters:
      - description: User id
        in: query
        name: user
        type: integer
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Size number
        in: query
        name: size
        type: integer
      - description: With the bearer started.
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.CoffeeShop'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ApiError'
      summary: Return liked coffee shops by user,
      tags:
      - likes
    post:
      consumes:
      - application/json
      description: Like a coffee shop
      parameters:
      - description: Like a coffee shop
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.LikeUnlikeCoffeeShopRequest'
      - description: With the bearer started.
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.LikeUnlikeCoffeeShopRequest'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ApiError'
      summary: Like a coffee shop
      tags:
      - likes
  /likes/{coffee_shop_id}:
    delete:
      consumes:
      - application/json
      description: Unlike a coffee shop
      parameters:
      - description: Unlike a coffee shop
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.LikeUnlikeCoffeeShopRequest'
      - description: Coffee shop id
        in: path
        name: coffee_shop_id
        required: true
        type: string
      - description: With the bearer started.
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.LikeUnlikeCoffeeShopRequest'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ApiError'
      summary: Unlike a coffee shop
      tags:
      - likes
  /login:
    post:
      consumes:
      - application/json
      description: Login a user using email and password receive a JWT as a response
        from a successful login
      parameters:
      - description: 'Login data: email and password'
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ApiError'
      summary: Login a user,
      tags:
      - users
  /signup:
    post:
      consumes:
      - application/json
      description: Register a user using email, username, password and password confirmation
      parameters:
      - description: 'Login data: email, password and password confirmation'
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.SignUpRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.EmptyBody'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ApiError'
      summary: Register a new user,
      tags:
      - users
  /users/{user_id}:
    delete:
      consumes:
      - application/json
      description: Delete the current user account
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      - description: With the bearer started.
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/models.EmptyBody'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ApiError'
      summary: Delete current user
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Get id, username, email, first name, last name and bio from a user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetUserResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/types.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ApiError'
      summary: Get an user account data,
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update the current user's bio, first name, last name and username
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      - description: 'User data: id, bio, firstName, lastName and username'
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateUserRequest'
      - description: With the bearer started.
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UpdateUserRequest'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ApiError'
      summary: Update current user,
      tags:
      - users
swagger: "2.0"